Move downlaod csv to vessel positions action
